feat(flush): Перенес флаш из сервисов в ручки
META: #91836
