+++ backend/src/Task/Command/Comment/Add/AddCommentOnTask.php
@@ -5,13 +5,11 @@ declare(strict_types=1);
 namespace App\Task\Command\Comment\Add;
 
 use App\Infrastructure\AsService;
-use App\Infrastructure\Flush;
 use App\Task\Domain\AddCommentToCompletedTaskException;
 use App\Task\Domain\Task;
 use App\Task\Domain\TaskComment;
 use App\Task\Domain\TaskCommentBody;
 use App\Task\Domain\TaskCommentId;
-use Psr\Log\LoggerInterface;
 
 /**
  * Хендлер добавления комментария к задаче
@@ -19,12 +17,6 @@ use Psr\Log\LoggerInterface;
 #[AsService]
 final class AddCommentOnTask
 {
-    public function __construct(
-        private readonly Flush $flush,
-        private readonly LoggerInterface $logger,
-    ) {
-    }
-
     /**
      * @throws AddCommentToCompletedTaskException
      */
@@ -32,13 +24,5 @@ final class AddCommentOnTask
     {
         $comment = new TaskComment($task, $commentId, new TaskCommentBody($command->commentBody));
         $task->addComment($comment);
-
-        ($this->flush)();
-
-        $this->logger->info('Задача прокомментирована', [
-            'taskId' => $task->getTaskId(),
-            'commentId' => $commentId->getValue(),
-            self::class => __FUNCTION__,
-        ]);
     }
 }
