+++ backend/src/Task/Http/CompleteTaskAction.php
@@ -5,10 +5,12 @@ declare(strict_types=1);
 namespace App\Task\Http;
 
 use App\Infrastructure\ApiException\ApiBadRequestException;
+use App\Infrastructure\Flush;
 use App\Infrastructure\SuccessResponse;
 use App\Task\Command\CompleteTask;
 use App\Task\Domain\Task;
 use App\Task\Domain\TaskAlreadyIsDoneException;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
@@ -21,14 +23,24 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 #[AsController]
 final class CompleteTaskAction
 {
-    public function __construct(private readonly CompleteTask $completeTask)
-    {
+    public function __construct(
+        private readonly CompleteTask $completeTask,
+        private readonly Flush $flush,
+        private readonly LoggerInterface $logger,
+    ) {
     }
 
     public function __invoke(Task $task): SuccessResponse
     {
         try {
             ($this->completeTask)($task);
+
+            ($this->flush)();
+
+            $this->logger->info('Задача завершена', [
+                'id' => $task->getTaskId(),
+                self::class => __FUNCTION__,
+            ]);
         } catch (TaskAlreadyIsDoneException $e) {
             throw new ApiBadRequestException($e->getMessage());
         }
