+++ backend/src/Task/Http/CreateTask/CreateTaskAction.php
@@ -5,11 +5,13 @@ declare(strict_types=1);
 namespace App\Task\Http\CreateTask;
 
 use App\Infrastructure\ApiException\ApiBadRequestException;
+use App\Infrastructure\Flush;
 use App\Task\Command\CreateTask\CreateTask;
 use App\Task\Command\CreateTask\CreateTaskCommand;
 use App\Task\Domain\TaskId;
 use App\User\SignUp\Domain\UserId;
 use InvalidArgumentException;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
@@ -22,8 +24,11 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 #[AsController]
 final class CreateTaskAction
 {
-    public function __construct(private readonly CreateTask $createTask)
-    {
+    public function __construct(
+        private readonly CreateTask $createTask,
+        private readonly Flush $flush,
+        private readonly LoggerInterface $logger,
+    ) {
     }
 
     public function __invoke(CreateTaskCommand $createTaskCommand, UserId $userId): TaskData
@@ -31,6 +36,13 @@ final class CreateTaskAction
         try {
             $taskId = new TaskId();
             ($this->createTask)($createTaskCommand, $taskId, $userId);
+
+            ($this->flush)();
+
+            $this->logger->info('Задача создана', [
+                'id' => $taskId,
+                self::class => __FUNCTION__,
+            ]);
         } catch (InvalidArgumentException $e) {
             throw new ApiBadRequestException($e->getMessage());
         }
