+++ backend/src/User/Profile/Http/ProfileSaveAction.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\User\Profile\Http;
 
 use App\Infrastructure\ApiException\ApiBadRequestException;
+use App\Infrastructure\Flush;
 use App\User\Profile\Command\SaveProfile\SaveProfile;
 use App\User\Profile\Command\SaveProfile\SaveProfileCommand;
 use App\User\Profile\Query\FindByUserId\FindProfileByUserId;
@@ -26,7 +27,8 @@ final class ProfileSaveAction
 {
     public function __construct(
         private readonly SaveProfile $saveProfile,
-        private readonly FindProfileByUserId $findProfileByUserId
+        private readonly FindProfileByUserId $findProfileByUserId,
+        private readonly Flush $flush,
     ) {
     }
 
@@ -34,6 +36,8 @@ final class ProfileSaveAction
     {
         try {
             ($this->saveProfile)($command, $userId);
+
+            ($this->flush)();
         } catch (InvalidArgumentException $e) {
             throw new ApiBadRequestException($e->getMessage());
         }
