+++ backend/src/User/SignIn/Command/CreateToken.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\User\SignIn\Command;
 
 use App\Infrastructure\AsService;
-use App\Infrastructure\Flush;
 use App\User\SignIn\Domain\UserToken;
 use App\User\SignIn\Domain\UserTokens;
 use App\User\SignUp\Domain\UserId;
@@ -17,7 +16,7 @@ use Symfony\Component\Uid\Uuid;
 #[AsService]
 final class CreateToken
 {
-    public function __construct(private readonly UserTokens $userTokens, private readonly Flush $flush)
+    public function __construct(private readonly UserTokens $userTokens)
     {
     }
 
@@ -25,7 +24,5 @@ final class CreateToken
     {
         $userToken = new UserToken($userTokenId, $userId);
         $this->userTokens->add($userToken);
-
-        ($this->flush)();
     }
 }
