+++ backend/src/User/SignIn/Command/DeleteToken.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\User\SignIn\Command;
 
 use App\Infrastructure\AsService;
-use App\Infrastructure\Flush;
 use App\User\SignIn\Domain\UserTokens;
 use Symfony\Component\Uid\Uuid;
 
@@ -15,10 +14,8 @@ use Symfony\Component\Uid\Uuid;
 #[AsService]
 final class DeleteToken
 {
-    public function __construct(
-        private readonly UserTokens $userTokens,
-        private readonly Flush $flush,
-    ) {
+    public function __construct(private readonly UserTokens $userTokens)
+    {
     }
 
     public function __invoke(Uuid $userTokenId): void
@@ -26,7 +23,5 @@ final class DeleteToken
         $userToken = $this->userTokens->getById($userTokenId);
 
         $this->userTokens->remove($userToken);
-
-        ($this->flush)();
     }
 }
