+++ backend/src/User/SignUp/Command/ConfirmEmail.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\User\SignUp\Command;
 
 use App\Infrastructure\AsService;
-use App\Infrastructure\Flush;
 use App\User\SignUp\Domain\EmailAlreadyIsConfirmedException;
 use App\User\SignUp\Domain\UserNotFoundException;
 use App\User\SignUp\Domain\Users;
@@ -17,7 +16,7 @@ use Symfony\Component\Uid\Uuid;
 #[AsService]
 final class ConfirmEmail
 {
-    public function __construct(private readonly Users $users, private readonly Flush $flush)
+    public function __construct(private readonly Users $users)
     {
     }
 
@@ -28,7 +27,5 @@ final class ConfirmEmail
     {
         $user = $this->users->findByConfirmToken($confirmToken);
         $user->confirm();
-
-        ($this->flush)();
     }
 }
